Time: 69 ms (24.91%) | Memory: 105.8 MB (36.90%) - LeetSync

+++ 2525-count-number-of-distinct-integers-after-reverse-operations/count-number-of-distinct-integers-after-reverse-operations.java
@@ -1,14 +1,14 @@
 class Solution {
     public int countDistinctIntegers(int[] nums) {
         int n = nums.length;
-        Set<Long> st = new HashSet<>();
+        Set<Integer> st = new HashSet<>();
         for (int i = 0; i < n; i++) {
-            long num = nums[i];
+            int num = nums[i];
             st.add(num);
             boolean valid = false;
-            long ans = 0;
+            int ans = 0;
             while (num != 0) {
-                long digit = num % 10;
+                int digit = num % 10;
                 ans = ans * 10 + digit;
                 num = num / 10;
             }
